+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- `average` method
 - `weighted_average` method
 - `weighted_kappa` method
 - `pycmAverageError` class